Create national Germany (DE) calendar alongside Xetra (XETR) market calendar
Repurposes HolidayCalendarServiceDE as a pure national calendar
containing exactly Germany's 9 official nationwide (bundesweite)
public holidays, confirmed accurate with no accuracy issues (unlike
CH's cantonal ambiguity). Introduces HolidayCalendarServiceXETR as a
verbatim, lossless copy of the former DE content (all 11 entries,
including Christmas Eve/New Year's Eve as FLOATING full closures with
weekend suppression) — both are confirmed genuinely NOT German public
holidays, correctly moved to the market-only class.

Moves DE's existing Christmas Eve/New Year's Eve collision tests to
XETR and renames the 30-year test accordingly.

Closes #240.
