+++ holiday-calendar-western/src/main/resources/META-INF/services/org.holiday.calendar.HolidayCalendarService
@@ -11,6 +11,7 @@ org.holiday.calendar.impl.HolidayCalendarServiceGBP
 org.holiday.calendar.impl.HolidayCalendarServiceUK
 org.holiday.calendar.impl.HolidayCalendarServiceUS
 org.holiday.calendar.impl.HolidayCalendarServiceUSD
+org.holiday.calendar.impl.HolidayCalendarServiceXETR
 org.holiday.calendar.impl.HolidayCalendarServiceXLON
 org.holiday.calendar.impl.HolidayCalendarServiceXNYS
 org.holiday.calendar.impl.HolidayCalendarServiceXTSE
