+++ holiday-calendar-western/src/test/java/org/holiday/calendar/impl/HolidayCalendarServiceDETest.java
@@ -18,7 +18,7 @@
 
 package org.holiday.calendar.impl;
 
-import org.holiday.calendar.HolidayCalendar;
+import org.holiday.calendar.HolidayCalendarService;
 import org.holiday.calendar.HolidayDate;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
@@ -31,9 +31,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertTrue;
 
 public class HolidayCalendarServiceDETest extends AbstractHolidayCalendarServiceTest {
 
@@ -48,9 +46,7 @@ public class HolidayCalendarServiceDETest extends AbstractHolidayCalendarService
     Iterator<Object[]> expectedHolidayNames() {
         final Object[] germanUnityDay = {"German Unity Day"};
         final Object[] labourDay = {"Labour Day"};
-        final Object[] christmasEve = {"Christmas Eve"};
-        final Object[] newYearsEve = {"New Year's Eve"};
-        return Arrays.asList(germanUnityDay, labourDay, christmasEve, newYearsEve).listIterator();
+        return Arrays.asList(germanUnityDay, labourDay).listIterator();
     }
 
     @DataProvider
@@ -68,45 +64,23 @@ public class HolidayCalendarServiceDETest extends AbstractHolidayCalendarService
         final Object[] boxingDay22 = {2022, "Boxing Day", LocalDate.of(2022, Month.DECEMBER, 26)};
         // Boxing Day 2023: Dec 26 is Tuesday -> no roll
         final Object[] boxingDay23 = {2023, "Boxing Day", LocalDate.of(2023, Month.DECEMBER, 26)};
-        // Christmas Eve 2024: Dec 24 is Tuesday -> present, unrolled (primary-sourced)
-        final Object[] christmasEve24 = {2024, "Christmas Eve", LocalDate.of(2024, Month.DECEMBER, 24)};
-        // New Year's Eve 2025: Dec 31 is Wednesday -> present, unrolled (primary-sourced)
-        final Object[] newYearsEve25 = {2025, "New Year's Eve", LocalDate.of(2025, Month.DECEMBER, 31)};
-        // Christmas Eve 2021: Dec 24 is Friday -> present, unrolled (collision year, see
-        // testChristmasEveAndChristmasDayCoexistOn24Dec2021 below)
-        final Object[] christmasEve21 = {2021, "Christmas Eve", LocalDate.of(2021, Month.DECEMBER, 24)};
         return Arrays.asList(christmas21, christmas22, christmas23,
-                             boxingDay21, boxingDay22, boxingDay23,
-                             christmasEve24, newYearsEve25, christmasEve21).listIterator();
+                             boxingDay21, boxingDay22, boxingDay23).listIterator();
     }
 
     @Test
-    public void testChristmasEveAndNewYearsEveOmittedOnWeekend() {
-        // 2023: Dec 24 and Dec 31 both fall on a Sunday, so Xetra/FWB simply doesn't
-        // list an exception day at all -- no shift, no entry.
-        final HolidayCalendar calendar = factory.create(CODE);
-        final Set<String> names = calendar.calculate(2023).stream()
-                .map(hd -> hd.getHoliday().getName())
-                .collect(Collectors.toSet());
-        assertFalse(names.contains("Christmas Eve"), "Christmas Eve must be omitted in 2023 (Sunday)");
-        assertFalse(names.contains("New Year's Eve"), "New Year's Eve must be omitted in 2023 (Sunday)");
-    }
-
-    @Test
-    public void testChristmasEveAndChristmasDayCoexistOn24Dec2021() {
-        // Dec 25, 2021 is a Saturday and rolls to Friday Dec 24 under DE's
-        // previousFridayOrFollowingMonday roll rule, landing on the same calendar date
-        // as the new non-rolling Christmas Eve holiday. Both facts are independently
-        // true and must both appear -- no accidental de-duplication.
-        final HolidayCalendar calendar = factory.create(CODE);
-        final List<HolidayDate> dec24 = calendar.calculate(2021).stream()
-                .filter(hd -> LocalDate.of(2021, Month.DECEMBER, 24).equals(hd.getDate()))
-                .toList();
-
-        assertEquals(dec24.size(), 2, "Expected both Christmas Day (rolled) and Christmas Eve on 2021-12-24");
-        final Set<String> names = dec24.stream().map(hd -> hd.getHoliday().getName()).collect(Collectors.toSet());
-        assertTrue(names.contains("Christmas Day"));
-        assertTrue(names.contains("Christmas Eve"));
+    public void testChristmasEveAndNewYearsEveAbsentFromCalculate() {
+        HolidayCalendarService service = factory.getService(CODE);
+        for (int year : List.of(2021, 2023, 2024, 2025)) {
+            List<HolidayDate> holidays = service.getHolidayCalendar().calculate(year);
+            Set<String> actualNames = holidays.stream()
+                    .map(hd -> hd.getHoliday().getName())
+                    .collect(Collectors.toSet());
+            assertFalse(actualNames.contains("Christmas Eve"),
+                    "Christmas Eve must not appear in calculate(" + year + ") — moved to XETR");
+            assertFalse(actualNames.contains("New Year's Eve"),
+                    "New Year's Eve must not appear in calculate(" + year + ") — moved to XETR");
+        }
     }
 
 }
